Change ints to specific uints

+++ prodos/block.go
@@ -14,7 +14,7 @@ import (
 )
 
 // ReadBlock reads a block from a ProDOS volume into a byte array
-func ReadBlock(reader io.ReaderAt, block int) ([]byte, error) {
+func ReadBlock(reader io.ReaderAt, block uint16) ([]byte, error) {
 	buffer := make([]byte, 512)
 
 	_, err := reader.ReadAt(buffer, int64(block)*512)
@@ -27,7 +27,7 @@ func ReadBlock(reader io.ReaderAt, block int) ([]byte, error) {
 }
 
 // WriteBlock writes a block to a ProDOS volume from a byte array
-func WriteBlock(writer io.WriterAt, block int, buffer []byte) error {
+func WriteBlock(writer io.WriterAt, block uint16, buffer []byte) error {
 	_, err := writer.WriteAt(buffer, int64(block)*512)
 	if err != nil {
 		errString := fmt.Sprintf("failed to write block %04X: %s", block, err.Error())
